+++ dfusion_rust_core/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 array-macro = "1.0.4"
 byteorder = "1.3.1"
 graph = { git = "https://github.com/graphprotocol/graph-node", tag = "v0.13.2" }
+mock-it = "0.2.0"
 mongodb = { git = "https://github.com/mongodb-labs/mongo-rust-driver-prototype", rev = "9203ec3216976cfbfddb07856b25d061965810af"}
 serde_derive = "1.0"
 serde = "1.0"
