+++ driver/src/mongo_db.rs
@@ -136,63 +136,3 @@ impl DbInterface for MongoDB {
         Ok(standing_orders)
     }
 }
-
-#[cfg(test)]
-pub mod tests {
-    use super::*;
-    use mock_it::Mock;
-
-    #[derive(Clone)]
-    pub struct DbInterfaceMock {
-        pub get_current_balances: Mock<H256, Result<models::AccountState, DatabaseError>>,
-        pub get_deposits_of_slot: Mock<u32, Result<Vec<models::PendingFlux>, DatabaseError>>,
-        pub get_withdraws_of_slot: Mock<u32, Result<Vec<models::PendingFlux>, DatabaseError>>,
-        pub get_orders_of_slot: Mock<u32, Result<Vec<models::Order>, DatabaseError>>,
-        pub get_standing_orders_of_slot: Mock<u32, Result<[models::StandingOrder; models::NUM_RESERVED_ACCOUNTS], DatabaseError>>,
-    }
-
-    impl DbInterfaceMock {
-        pub fn new() -> DbInterfaceMock {
-            DbInterfaceMock {
-                get_current_balances: Mock::new(Err(DatabaseError::new(Unknown, "Unexpected call to get_current_balances"))),
-                get_deposits_of_slot: Mock::new(Err(DatabaseError::new(Unknown, "Unexpected call to get_deposits_of_slot"))),
-                get_withdraws_of_slot: Mock::new(Err(DatabaseError::new(Unknown, "Unexpected call to get_withdraws_of_slot"))),
-                get_orders_of_slot: Mock::new(Err(DatabaseError::new(Unknown, "Unexpected call to get_withdraws_of_slot"))),
-                get_standing_orders_of_slot: Mock::new(Err(DatabaseError::new(Unknown, "Unexpected call to get_standing_orders_of_slot"))),
-            }
-        }
-    }
-
-    impl DbInterface for DbInterfaceMock {
-        fn get_current_balances(
-            &self,
-            current_state_root: &H256,
-        ) -> Result<models::AccountState, DatabaseError> {
-            self.get_current_balances.called(*current_state_root)  // https://github.com/intellij-rust/intellij-rust/issues/3164
-        }
-        fn get_deposits_of_slot(
-            &self,
-            slot: u32,
-        ) -> Result<Vec<models::PendingFlux>, DatabaseError> {
-            self.get_deposits_of_slot.called(slot)
-        }
-        fn get_withdraws_of_slot(
-            &self,
-            slot: u32,
-        ) -> Result<Vec<models::PendingFlux>, DatabaseError> {
-            self.get_withdraws_of_slot.called(slot)
-        }
-        fn get_orders_of_slot(
-            &self,
-            slot: u32,
-        ) -> Result<Vec<models::Order>, DatabaseError> {
-            self.get_orders_of_slot.called(slot)
-        }
-        fn get_standing_orders_of_slot(
-            &self,
-            slot: u32,
-        ) -> Result<[models::StandingOrder; models::NUM_RESERVED_ACCOUNTS], DatabaseError> {
-            self.get_standing_orders_of_slot.called(slot)
-        }
-    }
-}
